Decompose functions and data
